+++ voucher-stack/src/main/java/com/dwarfeng/voucher/stack/service/CheckerSupportMaintainService.java
@@ -1,7 +1,6 @@
 package com.dwarfeng.voucher.stack.service;
 
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
@@ -18,11 +17,4 @@ public interface CheckerSupportMaintainService extends BatchCrudService<StringId
 
     String ID_LIKE = "id_like";
     String LABEL_LIKE = "label_like";
-
-    /**
-     * 重置驱动器支持。
-     *
-     * @throws ServiceException 服务异常。
-     */
-    void reset() throws ServiceException;
 }