+++ CHANGELOG.md
@@ -32,7 +32,8 @@
 
 ### 功能移除
 
-- (无)
+- 去除支持实体维护服务的重置功能。
+  - com.dwarfeng.voucher.stack.service.CheckerSupportMaintainService。
 
 ---